+++ api/v1beta1/pipeline_types.go
@@ -242,7 +242,3 @@ type PipelineList struct {
 	metav1.ListMeta `json:"metadata,omitzero"`
 	Items           []Pipeline `json:"items"`
 }
-
-func init() {
-	SchemeBuilder.Register(&Pipeline{}, &PipelineList{})
-}